fix: fix misreadings
